+++ directorycomparewindow.h
@@ -39,7 +39,7 @@ class DirectoryCompareWindow : public QMainWindow
     Q_OBJECT
 
 public:
-    explicit DirectoryCompareWindow(const QString& pathLeft, const QString& pathRight, QWidget *parent = nullptr);
+    explicit DirectoryCompareWindow(const QString& pathLeft, const QString& pathRight, const CompareCaseSensitivity ccs, QWidget *parent = nullptr);
     ~DirectoryCompareWindow();
 
     /// Whether the left-side tree needs to be refreshed after the window closes.
@@ -48,7 +48,7 @@ public:
     /// Whether the right-side tree needs to be refreshed after the window closes.
     bool rightTreeNeedsUpdate() const;
 signals:
-    void compareDirectories(const QString& left, const QString& right);
+    void compareDirectories(const QString& left, const QString& right, const CompareCaseSensitivity ccs);
 
 protected:
     void closeEvent(QCloseEvent* event) override;
@@ -91,6 +91,7 @@ private:
     bool initialShowDone;
     QString leftPath;
     QString rightPath;
+    CompareCaseSensitivity compareCaseSensitivity;
     QThread thread;
 
     Compare* compare;